✨ Refactor ProductsToolbar to handle product creation and integrate householdId
